+++ lookup/filter_darwin.go
