+++ lookup/filter_linux.go
