+++ lookup/internal/lookup.go
@@ -1,5 +1,3 @@
-// +build darwin
-
 // Copyright © 2019 booster authors
 //
 // This program is free software: you can redistribute it and/or modify
@@ -20,28 +18,30 @@ package internal
 import (
 	"bytes"
 	"io"
+	"regexp"
 
-	"howett.net/plist"
+	"gopkg.in/pipe.v2"
 )
 
-// ExtractAppName is used to find the value of the "CFBundleExecutable" key.
-// "r" is expected to be an ".plist" encoded file.
-func ExtractAppName(r io.Reader) (string, error) {
-	rs, ok := r.(io.ReadSeeker)
-	if !ok {
-		var buf bytes.Buffer
-		if _, err := io.Copy(&buf, r); err != nil {
-			return "", err
-		}
-		rs = bytes.NewReader(buf.Bytes())
+type lsofDecoderFunc func(io.Reader) ([]*OpenFile, error)
+
+// OpenNetFiles uses ``lsof'' (or its platform dependent equivalent) to find
+// the list of open network files. It then filters the result using "rgx":
+// each line that does not match is discarded.
+func OpenNetFiles(rgx *regexp.Regexp) ([]*OpenFile, error) {
+	dec := lsofDecoder()
+
+	p := pipe.Line(
+		lsofCmd(),
+		pipe.Filter(func(line []byte) bool {
+			return rgx.Match(line)
+		}),
+	)
+	output, err := pipe.Output(p)
+	if err != nil {
+		return []*OpenFile{}, err
 	}
 
-	var data struct {
-		Name string `plist:"CFBundleExecutable"`
-	}
-	if err := plist.NewDecoder(rs).Decode(&data); err != nil {
-		return "", err
-	}
-
-	return data.Name, nil
+	buf := bytes.NewBuffer(output)
+	return dec(buf)
 }
